# artifacts/open_api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Chat APIs
  version: 0.215.37
  description: API for the Simple Chat, handling chats, conversations, documents, user settings, groups, feedback, prompts, models, safety, and admin functions.
servers:
  - url: /api
    description: API Root
...
paths:
...
  /userSearch:
    get:
      tags:
        - Users
      summary: Search for users via Microsoft Graph
      description: Searches for users in the organization's Azure Active Directory based on display name, email, or User Principal Name using the Microsoft Graph API. Requires user authentication and appropriate Graph API permissions for the application registration.
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true # Code requires a non-empty query
          schema:
            type: string
            minLength: 1
          description: The search term (prefix) to query against user display name, mail, or UPN.
      responses:
        '200':
          description: List of matching users (up to 10). Returns empty list if no matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSearchResult'
        '401':
          description: Unauthorized (Invalid or missing authentication token, or insufficient Graph API permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500': # Or other Graph API error codes (e.g., 403 if permissions missing)
          description: Internal Server Error or Microsoft Graph API request failed. Check details field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Graph API request failed"
                  details:
                    type: object # Or string depending on Graph error format
                    description: Detailed error information from Microsoft Graph.
  /user/settings:
    get:
      tags:
        - Users
        - Settings
      summary: Get current user's settings
      description: Retrieves the personalized settings document (including preferences like active group, layout) for the currently authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User settings retrieved successfully. Returns empty object or default structure if no settings document exists yet for the user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsDocument'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Failed to retrieve settings from database)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Users
        - Settings
      summary: Update current user's settings
      description: Updates specific fields within the authenticated user's settings document. Merges the provided settings with existing ones. Sets `last_updated` timestamp.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - settings
              properties:
                settings:
                  $ref: '#/components/schemas/UserSettings' # The inner settings object containing fields to update
      responses:
        '200':
          description: User settings updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User settings updated successfully"
        '400':
          description: Bad Request (Missing or invalid JSON body, 'settings' key missing or not an object)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error (Failed to read or update settings in database)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
...
    UserSearchResult:
      type: object
      description: Information about a user found via Graph API search.
      properties:
        id:
          type: string # Graph User ID (often UUID)
        displayName:
          type: string
        email:
          type: string
          format: email
          nullable: true # Email might not always be present depending on Graph object
    UserSettings: # Inner settings object used in GET response and POST/PATCH request
      type: object
      description: User-specific preference settings.
      properties:
        activeGroupOid:
          type: string
          format: uuid
          nullable: true
          description: ID of the group currently marked as active for the user.
        layoutPreference:
          type: string # e.g., 'default', 'compact', could be enum if values are fixed
          nullable: true
          description: User's preferred UI layout.
        splitSizesPreference:
          type: array
          items:
            type: number
            format: float # Or integer depending on how pane sizes are stored
          nullable: true
          description: User's preferred sizes for resizable UI panes.
        dockedSidebarHidden:
          type: boolean
          nullable: true
          description: User's preference for the docked sidebar visibility state.
        # Add other user-configurable settings keys here
    UserSettingsDocument: # The full document stored for a user in the database
      type: object
      description: The complete settings document for a user as stored in the database.
      properties:
        id:
          type: string # User ID (oid/sub), typically used as document ID
        user_id: # Can be redundant but often included for querying
          type: string
          description: User ID (oid/sub).
        settings:
          $ref: '#/components/schemas/UserSettings'
          description: The nested object containing user preferences.
        last_updated:
          type: string
          format: date-time
          description: Timestamp when the user settings were last modified.
        # Add any other top-level fields (like _ts for Cosmos DB)
  securitySchemes:
    bearerAuth: # Defines how API clients authenticate
      type: http
      scheme: bearer
      bearerFormat: JWT # Indicates the type of token expected
      description: "Azure AD issued JWT Bearer token required for authenticated endpoints. Include in the request header as 'Authorization: Bearer <YOUR_TOKEN>'."